fix for issue #2
The plugin was trying to set permissions on Simple Pages for students, without having the Simple Pages plugin installed. Added checks to only show options for Simple Pages when the plugin is installed, as well as before the ACL permissions are set. Finally, added Simple Pages to the list of optional plugins, so that it always runs its code after Simple Pages is set up